pyproject.toml
[tool.poetry]
name = "to-do-app"
version = "0.1.0"
description = ""
authors = ["TaleeBarbieri <talee.alexei@hotmail.com>"]
packages = [{include = "base", from='src'}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
django-environ = "^0.11.2"
requests = "^2.31.0"
psycopg2-binary = "^2.9.7"
django-admin-extra-buttons = "^1.5.6"
pillow = "^10.1.0"
natural-keys = "^2.1.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
